Write the we already uploaded file marker to the right location.
Due to an earlier hack I was writing the real uploaded marker & server
result files to the current directory, and not where the submitted
file was. (Because sometimes I was submitting to directories I didn't
have permissions for).

But this rule missed that, previously it worked because the
submitted_file_name was in the current directory so it was the same,
but with the subpool code I needed to match the hack I'd done earlier.